Increase error handling clarity
The .catch handler here is potentially confusing. If the fetch's .then() handler was modified by end-users to include code that raises runtime errors - anything from a simple typo to a call to a function that raises exceptions - those errors would be swallowed and the handler would report the app as being offline.